Working on invoke for invoking CMake.
